Port surface tests
